% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregator}
\alias{aggregator}
\title{Set the class of IDSR processed data to be able to merge countries in to one
dataset.}
\usage{
aggregator(x, output_path)
}
\arguments{
\item{x}{A data frame}

\item{output_path}{Path to where you would like to save a RDS file of merge data}
}
\description{
Set the class of IDSR processed data to be able to merge countries in to one
dataset.
}
\seealso{
\code{\link{clean_data}} for preparing IDSR data for use with threshold_checker,
\code{\link{missing_checker}} for missing flags and \code{\link{threshold_checker}}
for threshold flags.
}
