% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_data}
\alias{clean_data}
\title{Clean WHO IDSR processed data for use in weekly / monthly reports.
Renaming of diseases based on disease dictionary, setting threshold
type for disease based on threshold dictionary and creating an epiweek
variable.
Will add variables to your data set including: epiweek, threshold_type,
Alert_threshold and Recommendation (note the last three are placeholders to be
used for threshold_checker)}
\usage{
clean_data(x)
}
\arguments{
\item{x}{A data frame}
}
\description{
Clean WHO IDSR processed data for use in weekly / monthly reports.
Renaming of diseases based on disease dictionary, setting threshold
type for disease based on threshold dictionary and creating an epiweek
variable.
Will add variables to your data set including: epiweek, threshold_type,
Alert_threshold and Recommendation (note the last three are placeholders to be
used for threshold_checker)
}
