% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{country_feedback}
\alias{country_feedback}
\title{Create excel sheets with feedback for countries. Returns an excel workbook
for each country with one sheet showing missings and another showing threshold
flags, as well as recommendations for each.
Requires WHO IDSR data which has been prepared using the {clean_idsr} function,
and then run through the {missing_checker} and {threshold_checker} functions,
and all being stored in a list with a data frame for each country.}
\usage{
country_feedback(
  x,
  current_year = 2020,
  current_week = 1,
  output_path = here::here("Data Files", "Output")
)
}
\arguments{
\item{x}{A list of data frames which has been processed with {clean_idsr},
{missing_checker} and {threshold_checker}.}

\item{current_year}{The year of interest as numeric (e.g. the default is 2020)}

\item{current_week}{The week of interest as numeric. The default is 1.
Values less than 10 are padded with leading zeros, e.g. 1 becomes 01. (You
could also enter 01 and it would be fine)}

\item{output_path}{File path where you would like outputs to be saved to.
Within this folder a new folder will be created and named after the current week.
This defaults to your current directory, with subfolders of data > outputs > verification.
The year folder needs to exist already - only the current week folder will be created.}
}
\description{
Create excel sheets with feedback for countries. Returns an excel workbook
for each country with one sheet showing missings and another showing threshold
flags, as well as recommendations for each.
Requires WHO IDSR data which has been prepared using the {clean_idsr} function,
and then run through the {missing_checker} and {threshold_checker} functions,
and all being stored in a list with a data frame for each country.
}
\seealso{
\code{\link{clean_idsr}} for preparing IDSR data for use with threshold_checker,
\code{\link{missing_checker}} for missing flags and \code{\link{threshold_checker}}
for threshold flags.
}
