% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivarcolours.R
\name{make_colours}
\alias{make_colours}
\title{Create a combination of colours for plotting bivariate maps using {colorspace}}
\usage{
make_colours(og_pal1 = "OrRd", og_pal2 = "Blues", num1 = 5, num2 = 5)
}
\arguments{
\item{og_pal1}{A string indicated one of the available sequential palettes from
{colorspace}, the first value of this palette is used as the lowest in the combination.}

\item{og_pal2}{A string indicated one of the available sequential palettes from
{colorspace}}

\item{num1}{The number of colours to include in {og_pal1}, default is 5.}

\item{num2}{The number of colours to include in {og_pal2}, note that this does
not need to be the same as num1, but by default is.}
}
\description{
Create a combination of colours for plotting bivariate maps using {colorspace}
}
