% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missings.R
\name{missing_checker}
\alias{missing_checker}
\title{Flag weeks with missing counts for cases or deaths.
Requires WHO IDSR data which has been prepared using the {clean_data} function.
Will return variables for missing cases, missing deaths and a combined variable
with both in.}
\usage{
missing_checker(x)
}
\arguments{
\item{x}{A data frame}
}
\description{
Flag weeks with missing counts for cases or deaths.
Requires WHO IDSR data which has been prepared using the {clean_data} function.
Will return variables for missing cases, missing deaths and a combined variable
with both in.
}
\seealso{
\code{\link{clean_data}} for preparing IDSR data for use with missing_checker.
}
