% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholds.R
\name{threshold_checker}
\alias{threshold_checker}
\title{Flag weeks with case or death counts which are higher than the threshold defined
in threshold dictionary.
Requires WHO IDSR data which has been prepared using the {clean_idsr} function.
Will return variables for unique regions, alerts, and recommendations.}
\usage{
threshold_checker(x)
}
\arguments{
\item{x}{A data frame}
}
\description{
Flag weeks with case or death counts which are higher than the threshold defined
in threshold dictionary.
Requires WHO IDSR data which has been prepared using the {clean_idsr} function.
Will return variables for unique regions, alerts, and recommendations.
}
\seealso{
\code{\link{clean_idsr}} for preparing IDSR data for use with threshold_checker
}
