% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekreport.R
\name{week_report}
\alias{week_report}
\title{Wrapper for IDSR data processing and reporting functions in {epichecks}.
Will produce country feedback excel sheets (missings and thresholds),
country PDF letters, a weekly excel summary report and will aggregate country
data in to one large dataframe.}
\usage{
week_report(
  current_year = 2020,
  current_week = 1,
  week_start = "Monday",
  input_path = here::here("Data Files"),
  output_path = here::here("Data Files", "Output")
)
}
\arguments{
\item{current_year}{The year of interest as numeric (e.g. the default is 2020)}

\item{current_week}{The week of interest as numeric. The default is 1.
Values less than 10 are padded with leading zeros, e.g. 1 becomes 01. (You
could also enter 01 and it would be fine)}

\item{week_start}{First day of the week as accepted for {aweek}.
The default for this is "Monday" to confirm with ISO standards.}

\item{input_path}{File path where IDSR processed data is. The default is set
to your current directory, within the "Data Files" folder which should contain a
subfolder called "!Imported". Within "!Imported", there should be a folder for
each year - which has to fit to the current_year parameter above.
Within that a folder for each week, fitting to the current_week param.
All excel files within this folder will be read in.}

\item{output_path}{File path where you would like outputs to be saved to.
Within this folder a new folder will be created for the current year, and
within that the current week.
This defaults to your current directory, with subfolders of
Data Files > Output > 2020 > 01.
The year folder needs to exist already - only the current week folder will be created.
The file itself will labelled with the current week, e.g. "SummaryReport_2020-w01.xlsx"}
}
\description{
Wrapper for IDSR data processing and reporting functions in {epichecks}.
Will produce country feedback excel sheets (missings and thresholds),
country PDF letters, a weekly excel summary report and will aggregate country
data in to one large dataframe.
}
\seealso{
\code{\link{clean_idsr}} for preparing IDSR data for use with threshold_checker,
\code{\link{missing_checker}} for missing flags and \code{\link{threshold_checker}}
for threshold flags. In addition, see \code{\link{country_feedback}},
\code{\link{country_letters}} and \code{\link{weekly_summary}} for producing
feedback to countries; as well as \code{\link{aggregator}} to compile countries
in to one dataset.
}
