% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeksummary.R
\name{weekly_summary}
\alias{weekly_summary}
\title{Create an excel summary report of data submitted for week by each country.}
\usage{
weekly_summary(
  x,
  current_week = "2018-W35",
  output_path = here::here("Data", "Outputs", "Verification")
)
}
\arguments{
\item{x}{A data frame of IDSR data which has been created from {aggregator}}

\item{current_week}{The week of interest as a character or {aweek} object.
Needs to be in the correct format ("YYYY-Www").
The default "2018-W35" is just to demonstrate the necessary format.}

\item{output_path}{File path where you would like outputs to be saved to.
Within this folder a new folder will be created and named after the current week.
This defaults to your current directory, with subfolders of data > outputs > verification.
These folders need to exist already - only the current week folder will be created.
The file itself will labelled with the current week, e.g. "SummaryReport_2018-w35.xlsx"}
}
\description{
Create an excel summary report of data submitted for week by each country.
}
\seealso{
\code{\link{clean_data}} for preparing IDSR data for use with threshold_checker,
\code{\link{missing_checker}} for missing flags and \code{\link{threshold_checker}}
for threshold flags.
}
