% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeksummary.R
\name{weekly_summary}
\alias{weekly_summary}
\title{Create an excel summary report of data submitted for week by each country.}
\usage{
weekly_summary(
  x,
  current_year = 2020,
  current_week = 1,
  output_path = here::here("Data Files", "Output")
)
}
\arguments{
\item{x}{A data frame of IDSR data which has been created from {aggregator}}

\item{current_year}{The year of interest as numeric (e.g. the default is 2020)}

\item{current_week}{The week of interest as numeric. The default is 1.
Values less than 10 are padded with leading zeros, e.g. 1 becomes 01. (You
could also enter 01 and it would be fine)}

\item{output_path}{File path where you would like outputs to be saved to.
Within this folder a new folder will be created and named after the current week.
This defaults to your current directory, with subfolders of data > outputs > verification.
The year folder needs to exist already - only the current week folder will be created.}
}
\description{
Create an excel summary report of data submitted for week by each country.
}
\seealso{
\code{\link{clean_idsr}} for preparing IDSR data for use with threshold_checker,
\code{\link{missing_checker}} for missing flags and \code{\link{threshold_checker}}
for threshold flags.
}
